fixed nav links on mobile
Nav links sometimes didn't work, but neverfear, they now do. Also, i removed that large paragraph of dummy text, but that is something that still needs to be fixed: if an element is larger than the viewport height, it won't appear until i reach the bottom of the element. This can probably be fixed by making adjustments to the isInViewport() function.